Reorg code to move filtering out of native filewatcher
isWatched assumed that every file watcher knew everything that
was being watched. It is more efficient to assume that every
filewatcher knows what it *isn't* watching. The main FileWatcher
has the global knowledge to make the final decision on what is
being watched. This also seems to fix a bug where the native
file watcher would say it was watching everything, even the
items that were under a manual watch root.

Move the filtering logic for dirty files that are not under a
watch root to the main FileWatcher. The canonical file mapper
already has this data.

Switch back to a map version of the canonical file watcher (will
delete the tree implementation before final commit) because
filtering prevents storing sparse mapping data. We need to know
all roots to properly say if a file is under a root or not.

Change-Id: I14392be73b6b823fee9d4cda695cb4f86427e30d