Handle imprecise OS X file watcher
The OS X native file watcher may return the parent directory
when a file is changed. For example, if /foo/bar/a.txt is
changed, the file watcher may respond with DIRTY /foo/bar. These
situations require an additional flag to be passed back to the
master file watcher to indicate what path was changed AND if
the change was for that path of if the path being returned is a
parent of the path that actually had a change. This was encoded
in the "exact" parameter in the previous version of the
NativeFileWatcherImpl.